moved workers::confing and workers::input into csv_converter
