+++ csv_converter/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "csv_converter"
+version = "0.0.1"
+authors = ["Rob Rowe <robrowe04@gmail.com>"]
+
+[dependencies]
+serde = "1.0.70"
+serde_derive = "1.0.66"
+serde_json = "1.0.19"
+toml = "0.4"
