+++ csv_converter/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod config;
+pub mod input;
+
+#[macro_use]
+extern crate serde_derive;
+
+pub fn sound_check() {
+    println!("Is this thing on?");
+}
